+++ operator/internal/apply/auto/v1alpha1/workspacespec.go
@@ -35,6 +35,7 @@ type WorkspaceSpecApplyConfiguration struct {
 	Env                []v1.EnvVar                                `json:"env,omitempty"`
 	Resources          *v1.ResourceRequirements                   `json:"resources,omitempty"`
 	PodTemplate        *EmbeddedPodTemplateSpecApplyConfiguration `json:"podTemplate,omitempty"`
+	PulumiLogVerbosity *uint32                                    `json:"pulumiLogLevel,omitempty"`
 	Stacks             []WorkspaceStackApplyConfiguration         `json:"stacks,omitempty"`
 }
 
@@ -133,6 +134,14 @@ func (b *WorkspaceSpecApplyConfiguration) WithResources(value v1.ResourceRequire
 // If called multiple times, the PodTemplate field is set to the value of the last call.
 func (b *WorkspaceSpecApplyConfiguration) WithPodTemplate(value *EmbeddedPodTemplateSpecApplyConfiguration) *WorkspaceSpecApplyConfiguration {
 	b.PodTemplate = value
+	return b
+}
+
+// WithPulumiLogVerbosity sets the PulumiLogVerbosity field in the declarative configuration to the given value
+// and returns the receiver, so that objects can be built by chaining "With" function invocations.
+// If called multiple times, the PulumiLogVerbosity field is set to the value of the last call.
+func (b *WorkspaceSpecApplyConfiguration) WithPulumiLogVerbosity(value uint32) *WorkspaceSpecApplyConfiguration {
+	b.PulumiLogVerbosity = &value
 	return b
 }